Fix reactivity in table (#101)
* Fix reactivity in table
The headless table we use receives a Svelte `readable` or `writable` in
order to produce the table. The problem with using a `readable` is that
data can not be updated when it change after a request. Maybe I'm wrong
but I think the way is was done was not working properly. Now with a
writable we react each time data change and we produce a new array of
items to be passed to the headless table and everything reacts.

* In test check agains a fixed format instead of a Date object

* feature: use svelte's derive statements to fix table's reactivity

---------

Co-authored-by: Gerard Clos <clos.gerard@gmail.com>