feature: sync npm dependencies (#148)
Sync dependencies between root's package.json and latitude's server
package.json. This allows users to declare and install dependencies
with the regular npm flow. The main tradeoff is that conflicts between
latitude's server dependencies and the user dependencies might arise and
so we have to manually warn the user about them.