Fix StencilJS webcomponents build (#276)
We use StencilJS in our @latitude-data/webcomponents. StencilJS help use
author webcomponents and a thin wrapper in React. But in the future we
can add support for Vue or Svelte. The problem was that I develop the
iframe integration and it worked in developent but when I deployed a
production react example using our `LatitudeEmbed` component I realized
build was broken. This pr try to produce a valid Stencil + React build
for our component. We use `dist-custom-elements` as StencilJS target to
  be use together with the React wrapper generated by StencilJS