Improve app build time by not loading all the connectors (#298)
* Avoid loading all connectors in a Latitude app
We want to let users only install the connectors they need. This way
deploy and installation is more lightway

* minor: simplified some code and fixed bugs

I've tested it against a project that was using an old Latitude build
and it was breaking – so I'm making this a major change and updating the
changeset. I've also simplified the code a bit and fixed a couple of
bugs I've discovered with manual testing.

* minor: updated turbo

* feature: simplify dependency installs and improve logs

Stop using yellow color everywhere. Now dependencies are not synced with
the app and are simply installed in the project's root folder, the node
resolution algorithm will work it out.

---------

Co-authored-by: Gerard Clos <clos.gerard@gmail.com>