feature: update root dependencies on latitude update
Well, this is embarrasing... There's a dependency between latitude
server and connectors (obviously), and there's really no mechanism
to ensure compatibility between both. So, a user could install
version X of latitude server and install version Y of
a connector, and they wouldn't work together because there's no
mechanism to ensure compatibilty at install time.

For now, this commit implements an npm update of root packages on
`latitude update`. We are optimistically assuming most users will update
to the latest version so it's fair to assume an npm update will be the
wanted action most of the times.

This commit also:
- Implements a --force flag in update commands to override docker files
- Makes the update package add a package.json file if not present, same
  as setup command