Refactor compile query to pass SQL string from query file (#416)
We want to move the responsability of fining the query in the file
system from the connector to the source. After the last refactor now
this is easier to todo because we have access to the source and the
source manager in the connector.
With this change is easier to pass the the full path of the query and
look for relative paths and paths that lives outside the source of the
connector. We want to do both things in the upcoming changes