feature: CLI update command updates dockerfiles (#418)
Until now you had to pass a --force flag to update command in order to
update docker-related config files. This was done to protect users'
docker configs but in truth most people shouldn't be touching the docker
configuration and it's a pita to have to pass this flag everytime
there's an update.