Implement Materialize storage driver (#421)
When reading with DuckDB the parquet files from materialized queries we
need a way of knowing where those queries are stored. We want to allow
different ways of storing these parquet files. In this commit we start
implementing the driver and the disk implementation that will be used
almost always in development