Write a SQL query in Parquet file (#435)
We want to allow users to cache queries in a more permanent way. We're
doing persistent into parquet files. In this PR we introduce the
functionality for writing parquet files to the Source Manager. We also
introduce the concept of batched queries in our connectors. This means
now they can pull all the data from the query in a way that don't
exhaust the memory of the running machine. This can happen with huge
queries